requirements.txt
einops==0.8.0
flax==0.8.5
fvcore==0.1.5.post20221221
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.23.5
libsvm-official==3.35.0
lmdb==1.4.1
lpips==0.1.4
matplotlib==3.9.0
networkx==3.3
ninja==1.11.1.1
numpy==1.24.1
numpyro==0.15.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.5.22
orderedmultidict==1.0.1
pandas==2.2.2
pillow==10.3.0
pytorch-msssim==1.0.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
tensorboard==2.17.0
tensorboardX==2.6.2.2
timm==0.4.12
tokenizers==0.19.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==1.4.1
torchvision==0.14.1+cu117
tqdm==4.66.4
transformers==4.42.4
triton==2.3.1